requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
psycopg==3.2.10
psycopg-binary==3.2.10
sqlparse==0.5.3
